# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminUpdateUsersDTO:
    properties:
      githubProfile:
        maxLength: 30
        type: string
      name:
        type: string
      role:
        enum:
        - admin
        - user
        - ""
        type: string
      surname:
        type: string
      username:
        maxLength: 30
        type: string
    type: object
...
  dto.UpdatePasswordDTO:
    properties:
      confirmPassword:
        minLength: 8
        type: string
      newPassword:
        minLength: 8
        type: string
      password:
        type: string
    required:
    - confirmPassword
    - newPassword
    - password
    type: object
  dto.UpdateUserDTO:
    properties:
      githubProfile:
        maxLength: 30
        type: string
      name:
        type: string
      password:
        type: string
      surname:
        type: string
      username:
        maxLength: 30
        type: string
    required:
    - password
    type: object
  dto.UserDTO:
    properties:
      bestLanguage:
        type: string
      githubProfile:
        type: string
      name:
        type: string
      role:
        type: string
      surname:
        type: string
      username:
        type: string
    type: object
  response.BaseResponse:
    properties:
      data: {}
      data_count:
        type: integer
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost
info:
  contact: {}
  description: API Service for CodinLab
  title: API Service
  version: v1
paths:
  /private/admin/user:
    get:
      consumes:
      - application/json
      description: Retrieves All Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Get Users
      tags:
      - Admin
  /private/admin/user/{userID}:
    get:
      consumes:
      - application/json
      description: Retrieves User Profile
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Get Profile
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Updates User
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: New User Creds
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUsersDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Updates User
      tags:
      - Admin
...
  /private/user/:
    get:
      consumes:
      - application/json
      description: Gets users profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserDTO'
              type: object
      summary: GetProfile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates user
      parameters:
      - description: UpdateUser
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: UpdateUser
      tags:
      - User
  /private/user/password:
    put:
      consumes:
      - application/json
      description: Updates users password
      parameters:
      - description: UpdatePassword
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: UpdatePassword
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"